backend/requirements.txt
fastapi
uvicorn
pyserial
python-dotenv
SQLAlchemy>=2.0
alembic>=1.13
psycopg[binary]>=3.1
